+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Increased voiceovers version prediction error
+- Updated FPS Unlocker version which fixes new game version integration issue
+- Updated Turkish
+- Updated German
+- Updated Chinese
+- Updated Polish
 
 ## [3.9.2] - 13.11.2023